requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
fonttools==4.32.0
gdown==4.4.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.21.6
opencv-contrib-python==4.1.2.30
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
torch==1.8.2+cpu
torchvision==0.9.2+cpu
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
